Add files via upload
Added the backend code for Sudoku, including the functions that solves the puzzle, checks for empty sections within the puzzle, check if the input of the number within a certain section of the puzzle is valid (according to the Sudoku rules), and seeing if the solution is correct. Using this, we would soon later work on the actual javascript and the designing aspects of the Sudoku code.